requirements.txt
ethereum-etl
pandas
kafka-python
python-decouple
web3
ethereum-etl[streaming]
findspark
pyspark